requirements.txt
beautifulsoup4==4.11.2
cloudscraper==1.2.68
Flask==1.1.1
formatData==1.1
memory_profiler==0.61.0
numpy==1.19.5
pymongo==3.11.2
python-dotenv==1.0.0
pytz==2019.3
requests==2.23.0
